AWS-based-image-processing/requirements.txt
Flask==2.3.3
Flask-JWT-Extended==4.5.2
Pillow==10.0.0
boto3==1.28.62
botocore==1.31.62
gunicorn==21.2.0
uvicorn[standard]==0.23.2
requests==2.31.0
python-jose==3.3.0